+++ src/main/java/walkingkooka/spreadsheet/environment/SpreadsheetEnvironmentContextTesting.java
@@ -17,8 +17,12 @@
 
 package walkingkooka.spreadsheet.environment;
 
+import walkingkooka.environment.EnvironmentContext;
+import walkingkooka.environment.EnvironmentContextTesting;
+import walkingkooka.environment.EnvironmentContexts;
 import walkingkooka.net.AbsoluteUrl;
 import walkingkooka.net.Url;
+import walkingkooka.predicate.Predicates;
 import walkingkooka.spreadsheet.meta.SpreadsheetId;
 import walkingkooka.spreadsheet.net.HasSpreadsheetServerUrl;
 import walkingkooka.spreadsheet.net.HasSpreadsheetServerUrlTesting;
@@ -26,6 +30,7 @@ import walkingkooka.spreadsheet.storage.SpreadsheetStorageContext;
 import walkingkooka.storage.Storage;
 import walkingkooka.storage.StorageEnvironmentContextTesting;
 import walkingkooka.storage.StoragePath;
+import walkingkooka.storage.Storages;
 
 import java.util.Optional;
 
@@ -43,6 +48,30 @@ public interface SpreadsheetEnvironmentContextTesting extends StorageEnvironment
     AbsoluteUrl SERVER_URL = Url.parseAbsolute("https://example.com");
 
     SpreadsheetId SPREADSHEET_ID = SpreadsheetId.with(0x123);
+
+    Storage<SpreadsheetStorageContext> STORAGE = Storages.empty();
+
+    /**
+     * A {@link SpreadsheetEnvironmentContext} that contains {@link SpreadsheetEnvironmentContext#SERVER_URL} but not
+     * {@link SpreadsheetEnvironmentContext#SPREADSHEET_ID}.
+     */
+    SpreadsheetEnvironmentContext SPREADSHEET_ENVIRONMENT_CONTEXT = SpreadsheetEnvironmentContexts.basic(
+        STORAGE,
+        environmentContext()
+    );
+
+    private static EnvironmentContext environmentContext() {
+        final EnvironmentContext environmentContext = EnvironmentContextTesting.ENVIRONMENT_CONTEXT.cloneEnvironment();
+        environmentContext.setEnvironmentValue(
+            SpreadsheetEnvironmentContext.SERVER_URL,
+            SERVER_URL
+        );
+
+        return EnvironmentContexts.readOnly(
+            Predicates.always(), // all values are read-only
+            environmentContext
+        );
+    }
 
     // serverUrl........................................................................................................
