+++ src/main/java/walkingkooka/spreadsheet/validation/SpreadsheetValidatorContextTesting.java
@@ -18,6 +18,7 @@
 package walkingkooka.spreadsheet.validation;
 
 import org.junit.jupiter.api.Test;
+import walkingkooka.spreadsheet.environment.SpreadsheetEnvironmentContextTesting;
 import walkingkooka.spreadsheet.provider.SpreadsheetProviderContextTesting;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellReference;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelName;
@@ -27,7 +28,8 @@ import walkingkooka.spreadsheet.value.SpreadsheetErrorKind;
 import walkingkooka.validation.ValidatorContextTesting;
 
 public interface SpreadsheetValidatorContextTesting<C extends SpreadsheetValidatorContext> extends ValidatorContextTesting<C, SpreadsheetValidationReference>,
-    SpreadsheetProviderContextTesting<C> {
+    SpreadsheetProviderContextTesting<C>,
+    SpreadsheetEnvironmentContextTesting {
 
     @Test
     default void testValidationErrorWithCell() {
